Duration updated.

+++ ArmsModels/BaseModels/Operations/TripModel.cs
@@ -122,7 +122,7 @@ namespace ArmsModels.BaseModels
         public string EventName { get; set; }
         public string PlaceName { get; set; }
         public DateTime? EventDateTime { get; set; }
-        public int? EventDateTimeDiff { get; set; }
+        public decimal? EventDateTimeDiff { get; set; }
         public long? KMReading { get; set; }
         public long? KMReadingDiff { get; set; }
         public SharedModels.UserInfoModel UserInfo { get; set; }
