+++ ArmsServices/DataServices/Operations/TripService.cs
@@ -206,7 +206,7 @@ namespace ArmsServices.DataServices
                 EventName = reader.GetString("EventTypeName"),
                 PlaceName = reader.GetString("PlaceName"),
                 EventDateTime = reader?.GetDateTime("EventDateTime"),
-                EventDateTimeDiff = reader?.GetInt32("EventDateTimeDiff"),
+                EventDateTimeDiff = reader?.GetDecimal("EventDateTimeDiff"),
                 KMReading = reader?.GetInt64("KMReading"),
                 KMReadingDiff = reader?.GetInt64("KMReadingDiff"),
                 UserInfo = new ArmsModels.SharedModels.UserInfoModel
